subnet-calculator/api-fastapi-container-app/pyproject.toml
[project]
name = "subnet-calculator-api"
version = "1.0.0"
description = "FastAPI subnet calculator for Container Apps / App Service"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.118.0",
    "uvicorn[standard]>=0.34.0",  # ASGI server (replaces Azure Functions host)
    "pwdlib[argon2]>=0.2.0",      # Password hashing
    "pyjwt>=2.10.1",              # JWT authentication
    "pydantic>=2.10.3",           # Data validation
    "pydantic-settings>=2.6.1",   # Environment config
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.1",              # For FastAPI TestClient
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
